Refactor machinery into voxapi in order to minimize global namespace pollution.
